# openapi.yaml
openapi: 3.1.0
info:
  title: Vortex API
  version: 1.0.0
  description: |
    Local-first context refinery API for Vortex V1.

    **Product spec:** [idea.md](./idea.md)
    **Implementation spec:** [TECHNICAL.md](./TECHNICAL.md)

    Conventions:
    - All routes are space-scoped unless noted.
    - Long-running work (ingest, re-embed) returns `202 Accepted` with a `job_id`.
    - JSON field names use `snake_case` to align with the SQLite schema.
    - Timestamps are Unix milliseconds unless noted as ISO 8601 strings in export payloads.
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development (`pnpm dev`)

# V1 is local-first single-user; no auth layer.
security: []
...
paths:
...
  /api/spaces/{space_id}:
    parameters:
      - $ref: "#/components/parameters/space_id"
    get:
      tags: [Spaces]
      operationId: getSpace
      summary: Get space by ID
      responses:
        "200":
          description: Space detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpaceDetail"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Spaces]
      operationId: updateSpace
      summary: Update space metadata
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSpaceRequest"
      responses:
        "200":
          description: Updated space
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Space"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  parameters:
    space_id:
      name: space_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: State conflict or duplicate
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              examples: [validation_error, not_found, conflict, ingest_in_progress]
            message:
              type: string
            details:
              type: object
              additionalProperties: true
...
    Space:
      type: object
      required: [id, name, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        created_at:
          type: integer
          description: Unix ms
        updated_at:
          type: integer
          description: Unix ms
    SpaceDetail:
      allOf:
        - $ref: "#/components/schemas/Space"
        - type: object
          properties:
            counts:
              type: object
              properties:
                sources:
                  type: integer
                entities:
                  type: integer
                open_conflicts:
                  type: integer
                pending_review:
                  type: integer
...
    UpdateSpaceRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
...